Rename board.string to board.go_string
