+++ board/go_string.py
